feat: criteria based span counter
Adds a processor that creates a metric for spans count based on matching certain conditions.
Here's an example of the processor config:
```
hypertrace_spancounter:
  tenant_configs:
    - tenant_id: foo-bar-baz
      service_configs:
        - service_name: test-service
          span_configs:
            - label: test-service-attr-rules-spans
              span_attributes:
                - key: rpc.service
                  value: config.service.v1,ConfigService
                - key: rpc.method
                  value: GetAttributeRules
```
When the collector receives spans that match the tenant, service and span attributes above, it will add the count of those spans to the otelcol_criteria_span_count metric.